refactor: réorganiser les fichiers et mettre à jour les chemins d'inclusion
- Déplacer `fish.h` et `fish.cpp` vers le dossier `Entities`
- Déplacer `player.h` et `player.cpp` vers le dossier `Entities`
- Déplacer `shark.h` et `shark.cpp` vers le dossier `Entities`
- Mettre à jour les chemins d'inclusion dans les fichiers concernés
- Mettre à jour `CMakeLists.txt` pour refléter les nouveaux chemins des fichiers
