+++ CMakeLists.txt
@@ -26,21 +26,17 @@ add_executable(bloubloulespoissons main.cpp
         Entities/player.cpp
         Utility/env.h
         Utility/env.cpp
-        player.h
-        player.cpp
         menu.h
         menu.cpp
-        network/networking.cpp
-        network/networking.h
-        network/networking_client.cpp
-        network/networking_client.h
+        Network/networking.cpp
+        Network/networking.h
+        Network/networking_client.cpp
+        Network/networking_client.h
         Entities/shark.cpp
         Entities/shark.h
         Entities/fish.h
         Utility/display.cpp
         Utility/display.h
-        shark.cpp
-        shark.h
         Utility/utility.cpp
         Utility/utility.h
         Utility/close.cpp
